Gfx: Suppress int to float conversion warning

+++ Common/Gfx/CanvasD2D.cpp
@@ -32,7 +32,7 @@ D2D1_COLOR_F ToColorF(const Gdiplus::Color& color)
 
 D2D1_RECT_F ToRectF(const Gdiplus::Rect& rect)
 {
-	return D2D1::RectF(rect.X, rect.Y, rect.X + rect.Width, rect.Y + rect.Height);
+	return D2D1::RectF((FLOAT)rect.X, (FLOAT)rect.Y, (FLOAT)(rect.X + rect.Width), (FLOAT)(rect.Y + rect.Height));
 }
 
 D2D1_RECT_F ToRectF(const Gdiplus::RectF& rect)
